+++ webrtc/src/webrtc_c/api/video/video_frame_buffer.h
@@ -21,6 +21,14 @@ struct webrtc_VideoFrameBuffer_cbs {
   int (*width)(void* user_data);
   int (*height)(void* user_data);
   struct webrtc_I420Buffer_refcounted* (*ToI420)(void* user_data);
+  struct webrtc_VideoFrameBuffer_refcounted* (*CropAndScale)(struct webrtc_VideoFrameBuffer* self,
+                                                             int offset_x,
+                                                             int offset_y,
+                                                             int crop_width,
+                                                             int crop_height,
+                                                             int scaled_width,
+                                                             int scaled_height,
+                                                             void* user_data);
   void (*OnDestroy)(void* user_data);
 };
 
@@ -32,6 +40,26 @@ WEBRTC_EXPORT int webrtc_VideoFrameBuffer_height(
     const struct webrtc_VideoFrameBuffer* self);
 WEBRTC_EXPORT struct webrtc_I420Buffer_refcounted* webrtc_VideoFrameBuffer_ToI420(
     struct webrtc_VideoFrameBuffer* self);
+WEBRTC_EXPORT struct webrtc_VideoFrameBuffer_refcounted*
+webrtc_VideoFrameBuffer_DefaultCropAndScale(struct webrtc_VideoFrameBuffer* self,
+                                            int offset_x,
+                                            int offset_y,
+                                            int crop_width,
+                                            int crop_height,
+                                            int scaled_width,
+                                            int scaled_height);
+WEBRTC_EXPORT struct webrtc_VideoFrameBuffer_refcounted*
+webrtc_VideoFrameBuffer_CropAndScale(struct webrtc_VideoFrameBuffer* self,
+                                     int offset_x,
+                                     int offset_y,
+                                     int crop_width,
+                                     int crop_height,
+                                     int scaled_width,
+                                     int scaled_height);
+WEBRTC_EXPORT struct webrtc_VideoFrameBuffer_refcounted*
+webrtc_VideoFrameBuffer_Scale(struct webrtc_VideoFrameBuffer* self,
+                              int scaled_width,
+                              int scaled_height);
 WEBRTC_EXPORT struct webrtc_VideoFrameBuffer_refcounted*
 webrtc_VideoFrameBuffer_make_ref_counted(
     const struct webrtc_VideoFrameBuffer_cbs* cbs,